Renaming "function" files to "azure"
