+++ docs/examples/azure.md
