+++ src/index.ts
@@ -14,7 +14,7 @@ import DiscordHTTPError from './errors/DiscordHTTPError';
 import DiscordRESTError from './errors/DiscordRESTError';
 
 import AWSLambdaServer from './servers/lambda';
-import AzureFunctionServer from './servers/function';
+import AzureFunctionServer from './servers/azure';
 import ExpressServer from './servers/express';
 import FastifyServer from './servers/fastify';
 import GatewayServer from './servers/gateway';
